requirements.txt
telethon
python-dotenv
requests
asyncio